backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
geoalchemy2==0.15.2
httpx==0.28.1
feedparser==6.0.11
beautifulsoup4==4.12.3
spacy==3.8.3
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic==2.10.3
anthropic==0.40.0
geopy==2.4.1
lxml==5.3.0
